chat import now sanitizes also
