+++ tgui/packages/tgui-panel/settings/settingsImExport.ts
@@ -1,6 +1,7 @@
 import { omit } from 'es-toolkit';
-import { chatPagesRecordAtom } from '../chat/atom';
-import { importChatState } from '../chat/helpers';
+import { chatPagesRecordAtom, mainPage } from '../chat/atom';
+import { startChatStateMigration } from '../chat/migration';
+import type { Page, StoredChatSettings } from '../chat/types';
 import { store } from '../events/store';
 import { storedSettingsAtom } from './atoms';
 import { startSettingsMigration } from './migration';
@@ -53,8 +54,39 @@ export function importChatSettings(settings: string | string[]): void {
   const settingsPart = omit(ourImport, ['chatPages']);
 
   if ('chatPages' in ourImport && ourImport.chatPages) {
-    importChatState(ourImport.chatPages);
+    const chatPart = rebuildChatState(ourImport.chatPages);
+    if (chatPart) {
+      startChatStateMigration(chatPart);
+    }
   }
 
   startSettingsMigration(settingsPart);
+}
+
+/** Reconstructs chat settings from just the record */
+function rebuildChatState(
+  pageRecord: Record<string, Page>,
+): StoredChatSettings | undefined {
+  const newPageIds: string[] = Object.keys(pageRecord);
+  if (!newPageIds) return;
+
+  // Correct any missing keys from the import
+  const merged: Record<string, Page> = { ...pageRecord };
+  for (const page of newPageIds) {
+    merged[page] = {
+      ...mainPage,
+      ...pageRecord[page],
+      unreadCount: 0,
+    };
+  }
+
+  const rebuiltState: StoredChatSettings = {
+    version: 1,
+    scrollTracking: true,
+    currentPageId: newPageIds[0],
+    pages: newPageIds,
+    pageById: merged,
+  };
+
+  return rebuiltState;
 }
